requirements.txt
aiogram==3.13.1
google-api-python-client==2.151.0
google-auth==2.35.0
APScheduler==3.10.4
pydantic==2.9.2
pydantic-settings==2.6.1
portalocker==2.10.1
python-json-logger==2.0.7
